[M]fix housing item type

+++ saint-coinach/src/main/java/ffxiv/housim/saintcoinach/xiv/entity/housing/HousingItem.java
@@ -11,7 +11,7 @@ public abstract class HousingItem extends XivRow {
     }
 
     public int getModelKey() {
-        return asInt32("ModelKey");
+        return asInt16("ModelKey");
     }
 
     public Item getItem() {
